Basic code cleanup and fixing the bug in IsImagePath so it checks the path for the suffix, and not the suffix for the suffix.
