+++ src/Imageflow.Net.Server/ImageflowMiddlewareExtensions.cs
@@ -1,8 +1,5 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Imageflow.Server
 {
