+++ src/Imageflow.Net.Server/Structs/ImageData.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Imageflow.Server.Structs
+{
+    public struct ImageData
+    {
+        public ArraySegment<byte> resultBytes;
+        public string contentType;
+    }
+}
