+++ src/Imageflow.Net.Server/Structs/ResizeParams.cs
@@ -1,0 +1,13 @@
+﻿namespace Imageflow.Server.Structs
+{
+    public struct ResizeParams
+    {
+        public bool hasParams;
+        public string commandString;
+
+        public override string ToString()
+        {
+            return commandString;
+        }
+    }
+}
