+++ src/Imageflow.Net.ServerHelpers/Class1.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Imageflow.Net.ServerHelpers
+﻿namespace Imageflow.Net.ServerHelpers
 {
     public class Class1
     {
